Sacado Hardvok xD

+++ myGame/Engine/Mathjavi.cpp
@@ -25,10 +25,10 @@ void MATHJAVI::eulerToQuaternion(float fRotX, float fRotY, float fRotZ, float& o
 	double c1c2 = c1*c2;
 	double s1s2 = s1*s2;
 
-	orQW = c1c2*c3 - s1s2*s3;
-	orQX = c1c2*s3 + s1s2*c3;
-	orQY = s1*c2*c3 + c1*s2*s3;
-	orQZ = c1*s2*c3 - s1*c2*s3;
+	orQW = (float)(c1c2*c3 - s1s2*s3);
+	orQX = (float)(c1c2*s3 + s1s2*c3);
+	orQY = (float)(s1*c2*c3 + c1*s2*s3);
+	orQZ = (float)(c1*s2*c3 - s1*c2*s3);
 }
 
 void MATHJAVI::quaternionToEuler(float qX, float qY, float qZ, float qW, float& orfRotX, float& orfRotY, float& orfRotZ)
